+++ python_modules/dagit/dagit/schema/__init__.py
@@ -7,6 +7,6 @@ dauphin = DauphinRegistry()
 def create_schema():
     # dauphin depends on import-based side effects
     # pylint: disable=W0611
-    from dagit.schema import generic, roots, pipelines, execution, runs, errors
+    from dagit.schema import generic, roots, pipelines, execution, runs, errors, legacy_types, types
 
     return dauphin.create_schema()
